requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
awscli==1.40.20; python_version >= '3.9'
awsebcli==3.24.1
bcrypt==4.3.0; python_version >= '3.8'
blessed==1.21.0; python_version >= '2.7'
botocore==1.38.21; python_version >= '3.9'
cement==2.10.14
certifi==2025.4.26; python_version >= '3.6'
cffi==1.17.1; python_version >= '3.8'
charset-normalizer==3.4.2; python_version >= '3.7'
cloudinary==1.44.0
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
cryptography==45.0.2; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
decorator==5.2.1; python_version >= '3.8'
deprecated==1.2.18; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
django==5.2.1; python_version >= '3.10'
django-api==0.1.3
django-cors-headers==4.7.0; python_version >= '3.9'
djangorestframework==3.16.0; python_version >= '3.9'
docutils==0.19; python_version >= '3.7'
fabric==3.2.2
idna==3.10; python_version >= '3.6'
invoke==2.2.0; python_version >= '3.6'
jmespath==1.0.1; python_version >= '3.7'
packaging==24.2; python_version >= '3.8'
paramiko==3.5.1; python_version >= '3.6'
pathspec==0.12.1; python_version >= '3.8'
psycopg2-binary==2.9.11; python_version >= '3.9'
pyasn1==0.6.1; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pymysql==1.1.0; python_version >= '3.7'
pynacl==1.5.0; python_version >= '3.6'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.1.0; python_version >= '3.9'
pyyaml==6.0.2; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
rsa==4.7.2; python_version >= '3.5' and python_version < '4'
s3transfer==0.12.0; python_version >= '3.9'
semantic-version==2.10.0; python_version >= '2.7'
setuptools==80.8.0; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlparse==0.5.3; python_version >= '3.8'
stripe==12.1.0; python_version >= '3.6'
termcolor==2.5.0; python_version >= '3.9'
typing-extensions==4.13.2; python_version >= '3.8'
urllib3==1.26.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
wcwidth==0.2.13
wrapt==1.17.2; python_version >= '3.8'
